Fix possible infinite loop reading CHM archives
According to the mspack documentation:
  [The mspack read() callback function should] return the number of bytes
  successfully read (this can be less than the number requested), zero to mark
  the end of file, or less than zero to indicate an error.
  The library does not "retry" reads and assumes short reads are due to
  EOF, so you should avoid returning short reads because of transient errors.

Our implementation returns the number of bytes read, as required, but
fails to update the internal offset when reading from an fmap. This
means that if mspack does retry a partial read, it will return the same
bytes in perpetuity.

This commit updates the offset for partial reads.

It also appears as though our implementation would return more bytes
than read if a partial read occurs if using a file descriptor instead of
an fmap. I've corrected this to return the number of bytes read.

Thank you to Michał Dardas for reporting this issue.