GH-actions: remove CIFuzz workflow
The CIFuzz tool was observed detecting an issue in code unrelated to a
pull request. While it was relatively benign in this case, it has the
potential to accidentally disclose a vulnerability publicly without
allowing a non-disclosure period for us to publish a fixed patch
version.

I'm not comfortable with this risk, so I'm removing the workflow.