Fix NULL param crash when caching
Since converting the hash variable from a stack array to a pointer, the
pointer may now be NULL if the file is truncated after the scan starts
but before the hash is calculated. This race condition would result in
a NULL pointer dereference and crash.

This commit adds additional NULL parameter checks.

Thanks to Alexander Patrakov and Antoine Gatineau for reporting this issue.

Resolves: https://github.com/Cisco-Talos/clamav/issues/440