Windows: Fix crash in Debug builds when scanning some PE files
The endianness conversion when reading the PE image import descriptor is
making the change in-place in the fmap.
On Windows, the fmap is read-only and so in Debug builds that's causing
a crash.

This uses a buffer on the stack and copies each image import descriptor
before doing the conversions and then processing each.