Ctest, Valgrind: disable possibly-lost alerts
We've observed a number of false positives with iconv, pthread, libxml2,
etc regarding memory allocation on init that is never cleaned up but
isn't a leak. Now we're seeing a whole tonne more from Rust libraries
with wild stack traces that couldn't be easily suppress.

So the solution is to disable any alerts from possible-leaks and only
show definite-leaks.