Rust utils: Add C-Rust FFI error module
The new frs_error module makes it so you can painlessly transfer Rust
errors to-and-from C.

On the Rust-side, on failure it returns an error struct to the C caller.
The C caller detects the failure by checking the return value.
On failure, the caller may use an additional frs_error Rust API to get a
formatted error message.

Two variants are provided:
- One for functions that return a boolean and have an out-parameter for
  function output.
- One for functions that return a pointer, where NULL means failure and
  non-NULL contains the successful function output.