PE parser: fix recently introduced NULL dereference crash
Commit f82492aef4e6b7b30e892a298e0d799ae99ef684 fixed a crash in Windows
debug builds but in so doing accidentally introduced a possible crash
when scanning PE files that lack import tables.  The issue being that
the openssl hashing functions try to "finish" a hash that was never
started.

This commit fixes the issue by returning CL_BREAK instead of CL_SUCCESS
when the import table doesn't exist or RVA is invalid so that we can
differentiate between successfully calculating the hashes and
successfully skipping the hashing process.