libclamav: Fix benign uninitialized variable used
In the core `cli_magic_scan()` function, early termination of the scan
before hashing the file, such as a return from the prescan-callback,
could result in the hashed_size variable being uninitalized when calling
into `cache_add()`. That's okay because we now skip caching of the md5 hash
when it is NULL, hence "benign". This is just a minor code quality fix.