Free correct variable in signature load error handling
We don't allocate a copy of the signature name to store in the AC
pattern structure for logical signature patterns because it is already
stored in the logical signature structure. But oss-fuzz found that we're
freeing that virname in when an error happens even if it wasn't copied.

This fix checks the allocation before MPOOL_FREE.

Since virname is passed in, and only cloned under certain condtions,
check to see that it has actually been cloned before freeing it in any
cleanup code.

Resolves: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=45205