1611 - add bounds checking to wdb pattern buffer
During wdb load, it was possible to go beyond the bounds
of the pattern buffer due to two subsequent increment ops
with no bounds checking in between.

This issue was reported by external researchers and
they provided the fix as well.

Based on our own research, this is a defect but not a vulnerability.