XLS parser: Fix issue where alert is lost
It is possible when not using allmatch mode that an alert on an XLS file
may be "lost". That is, it first reports as FOUND but then later reports
as OK.

The issue is essentially that the way allmatch and alert reporting is
done, it is easy to accidentally change the return value at one
intermediate layer and forget about the alert.

This fix doesn't clean up the design flaw, but does resolve this
specific bug.

Resolves:
- https://github.com/Cisco-Talos/clamav/issues/442
- https://github.com/Cisco-Talos/clamav/issues/521