Fix possible 2-byte overread when adding sig pattern
It is possible to create a signature pattern that tries to add a
zero-byte matching pattern to the A-C trie. A missing check at this
stage can end up with a 2-byte overread when indexing the (empty)
pattern to make sure the bytes added to the A-C trie are static and
not both zero.

This over read issue is not a vulnerability.

This commit fixes the issue by adding a check for the pattern length.

Resolves: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=43832

Also added:
- type casts and a "fall-through" comment to silence compile warnings.
- a few additional length checks to protect against an additional 1-byte
over read.