Fix scan error when image fuzzy hash fails
Scanning a file containing an image where it fails to calculate the
image fuzzy hash may cause the entire scan to end with an error.

This commit fixes that by ignoring any image fuzzy hash failure.
If calculating the hash fails, the hash is left uninitialized (all
zeroes). To prevent doing image fuzzy hash lookups when it failed,
I added a boolean that is set to true only after creating the hash.

Fixes: https://github.com/Cisco-Talos/clamav/issues/600
Fixes: https://github.com/Cisco-Talos/clamav/issues/593