Fix bug with logical signature Intermediates feature
The file type check for the logical signature Intermediates feature is
off-by-one and so the first layer checked is supposed to be the first
parent of the current layer, but is in fact the current layer.

This commit fixes the issue by correctly checking starting at -2 instead
of -1. (-1 is the current layer, 0 is the outermost layer).