Tests: allmatch with pattern match + embedded file type recognition
There is a bug where because embedded file type recognition is during
the raw-scan (i.e. pattern match of a given layer), if it finds a
malware sig match in the raw-scan it will lose the metadata about
embedded file types found, and will fail to extract them.
In all-match mode, we want it to keep extracting and scanning despite
matches, so this is a bug.

This test verifies that it will match on a signature, and also find the
embedded zip, extract it, and match on the stuff within.