Fix image parser match reporting issue
When using --heuristic-precendence, the heuristic alerts in the image
parsers much be reported as regular matches, which means that when not
in all-match mode, the scan should abort. Right now, those matches
aren't being recorded so scanning continues in archives containing more
than just the malformed image file.

This fix returns the "CL_VIRUS" status code when heuristic-precendence
is enabled and allmatch is not enabled.