Fix issue reporting all Heuristic/PUA matches in allmatch mode
When potentially unwanted indicators are found, they are not reported
right away, except in heuristic-precedence mode.  At the end of the
scan, we then report all of the heuristic/PUA alerts.  In my initial
overhaul of the allmatch mode, I introduced an issue where it is never
reporting more than 1 heuristic/PUA alert.

This commit fixes that by reporting all potentially-unwanted indicators
at the end. Note that instead of using `cli_virus_found_cb()` which
would report the top layer file descriptor as well in the callback,
I'm reporting `-1` for the file descriptor instead, because we don't
know when the alert was added.  If it was at a deeper layer in the file,
we would not longer have access to it at the end of the scan.

Also removed excessive debug statements '... infected with ...' that use
`cli_get_last_virus()` because it doesn't really reflect all of the
matches in allmatch mode, because the match would've been reported
previously (so it is excessive), and because I don't like the word
'infected'. :)