Matcher: Remove allmatch checks and significantly tidy code
Significantly tidy the `cli_scan_fmap()` function, and add comments to
explain how it all works.

Add SHA1 and SHA256 digest variables to the FMAP structure in addition
to the existing MD5. Add a function to set the hash so that when we
calculate the hashes for hash matching, we save them for subsequent FP
matching. This enabled me to remove the extra "hash-only" FP check from
`cli_scan_fmap()`. This will also make it easier to switch the clean
cache hash algorithm to SHA256 in the future.

Remove extra allmatch checks that are no longer required.

Add a new header to prevent #include order issues.