Fix PE pattern bounding (S#) for simple pattern matches
Simple pattern matches that don't use fancier features like wildcards
use the Boyer-Moore pattern matcher. Unfortunately, relative offsets
such as PE section bounding is not fully implemented for our Boyer-Moore
matcher. Specifically, the minimum offset is set but not the maximum,
which would result in matching outside of the given section. But in
addition, the `bm_shift` mechanism used in `cli_bm_scanbuff()` doesn't
work with these relative offsets and will never even attempt to do the
match.

This commit fixes the issue by detecting the 'S' in the `offset`
variable at signature load time and assigning the pattern to the
Aho-Corasick matcher, where everything appears to work as intended.