Tests: unit tests for cl_load(), cl_cvdverify(), cl_cvdunpack()
Some basic testing is needed for the new cl_cvdunpack() API, so this
commit adds basic unit tests for that.

For reasons unknown, a number of cl_* API's have stubs for unit tests
that weren't filled out.  The CVD load/verify ones in particular
required access to a signed CVD.  We actually ship a very basic signed
CVD with the databases now, so I added tests for those while I was at it.