Fix caching bug with Intermediates & Container conditions
Signatures that have conditions on a Container type or Intermediates
types may not match if the target file is scanned outside of the
container first, and cached as clean.

The solution to this is to disable caching for layers that would
have matched except for the Container (or Intermediates) condition.