Set -rc version suffix; bump libclamav SO version
In preparation for the release candidate, changing the version suffix
from devel to rc.

Increment SO current version to 11, because we added to the public API.
Increment SO age to 2, because we didn't change anything or remove
anything from the public API.

Added functionality level values to the bytecode API enum for recent
patch versions.