Fix memory compare length in OLE2 AES key verification
Coverity found two slightly different buffer length issues
with the memcmp at the end of the verify_key_aes() function.

The problem is that the hash must be a SHA1, but the the verifier hash
field is up to 32 bytes, and the structure may also lie about the size
of the hash, saying it's more than 32 bytes.
Realistically, it should be 20 bytes if it is functioning correctly,
so we shouldn't really care about the extra 12 bytes.