Freshclam: fix incremental update on CLD database
When adding the `cl_cvdunpack()` API that (optionally) verifies the
database signature, we used it in libfreshclam in a place where it may
also unpack CLD database archives. CLD's may not be verified, because
the signature information is no longer valid after incremental update.

This commit fixes the issue by only verifying the unpack if the file is
a CVD and not a CLD.