CMake: remove option to build using external TomsFastMath library
ClamAV requires non-default build options for TomsFastMath to support
bigger floating point numbers. Without this change, database and
Windows EXE/DLL authenticode certificate validation may fail.

The option to build with an external TomsFastMath library should be
removed. The `ENABLE_EXTERNAL_TOMSFASTMATH` build should be ignored,
with some sort of warning.

This commit removes the option. If used, the build will not fail, but
CMake will print a warning that the unused option was ignored.