OLE2: Sector size check before checking for encryption
This bug was introduced during 1.0 development and does not affect any
stable release.

An OLE2 document that claims to have a sector size (big block size) over
28 will cause a signed integer overflow for our local off_t variable,
when multiplied by 4. This will then result in an invalid read and
segfault.

The fix is two-fold:
1. Make sure the sector size (big block size) is not greater
   than 28, so we don't accidentaly overflow an int32_t.
2. Perform the size check _before_ using the variable.

This commit also fixes some variable types from `unsigned int` or
`off_t` over to the more appropriate `size_t`, and from `int` or
`uint32_t`, over to `bool`.

Fixes: https://bugs.chromium.org/p/oss-fuzz/issues/detail?id=52639