Fix error when loading some WDB databases
An error occurs when loading the WDB database with this line:

  X:urldefense\.com:.+

This happens because it goes to add a new suffix for the second domain
but there is nothing there, so it attempts to undo by shrinking the
suffix list back to original size. The suffix list in this case goes
back to zero, and our custom CLI_REALLOC function balks at reallocing to
zero.

I think the correct solution is to not free if it goes to zero, just in
case it is dereferenced later. It at least doesn't leak, so this is a
safe solution.