Add a version script for libclamav and libfreshclam
Without a version script all symbols will be exported which are public
within the libclamav library. This is true for the officially exported
symbols as well as all the public symbols which are used within
libclamav such as the rust symbols or the included libs like 7z and its
XZ_SIG symbol.

There is already a .map preset to limit the exported symbols of
libclamav and libfreshclam. Use it.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>