Add a version script for libclamunrar and libclamunrar_iface.
Without a version script all symbols will be exported which are public
within the libclamunrar* libraries. This is true for the officially exported
symbols as well as all the public symbols which are used within
libraries and are not inteded for export.

There is already a .map preset to limit the exported symbols of
libclamunrar and libclamunrar_iface. Use it.

Signed-off-by: Sebastian Andrzej Siewior <sebastian@breakpoint.cc>