Test: Check that a trusting a cert won't trust whole archive
This test scans a ZIP containing:
1. An executable signed with a trusted certificate
2. A file that will alert because of a signature match

The goal is to verify that the scan isn't terminated early and that
the entire archive isn't trusted simply because the inner executable is
trusted.
That is, that the "CL_VERIFIED" status does not propagate up out of the
magic_scan of a specific layer.